Split searcher.py
