+++ syne_tune/optimizer/schedulers/searchers/__init__.py
@@ -16,10 +16,12 @@ from syne_tune.try_import import try_import_gpsearchers_message
 
 from syne_tune.optimizer.schedulers.searchers.searcher import (  # noqa: F401
     BaseSearcher,
-    SearcherWithRandomSeed,
     impute_points_to_evaluate,
-    extract_random_seed,
+)
+from syne_tune.optimizer.schedulers.searchers.searcher_base import (  # noqa: F401
+    SearcherWithRandomSeed,
     SearcherWithRandomSeedAndFilterDuplicates,
+    extract_random_seed,
 )
 from syne_tune.optimizer.schedulers.searchers.random_grid_searcher import (  # noqa: F401
     RandomSearcher,
@@ -31,10 +33,7 @@ from syne_tune.optimizer.schedulers.searchers.searcher_factory import (  # noqa:
 
 __all__ = [
     "BaseSearcher",
-    "SearcherWithRandomSeed",
     "impute_points_to_evaluate",
-    "extract_random_seed",
-    "SearcherWithRandomSeedAndFilterDuplicates",
     "RandomSearcher",
     "GridSearcher",
     "searcher_factory",
